feat: refactor workflow to use GraphWorkflowHelper and update imports
